Allow loading static content in the Irydium viewer
The directory to serve is specified via a command line argument
(defaulting to the current directory). This has a number of nice
side-effects:

1. You can load components or data stored in the current directory
  when doing a local workflow.
2. You can load components when running the viewer against a notebook
  in static via `npm run dev -- <path-to-notebook>`. This is *very*
  helpful for local development workflows (previously other components
  wouldn't import at all, and the only way to get around it was to
  modify the notebook to point only to absolute URLs e.g. on
  irydium.dev)
