+++ .changeset/warm-carpets-pay.md
@@ -1,0 +1,6 @@
+---
+"@irydium/compiler": patch
+"@irydium/viewer": patch
+---
+
+Allow loading static content in the Irydium viewer
