+++ package.json
@@ -18,7 +18,7 @@
     "test": "jest",
     "format": "prettier --plugin-search-dir=. --write .",
     "format:check": "prettier --plugin-search-dir=. --check .",
-		"release": "pnpm -r build && changeset publish"
+    "release": "pnpm -r build && changeset publish"
   },
   "devDependencies": {
     "@babel/core": "^7.15.5",
@@ -48,7 +48,6 @@
     "@typescript-eslint/eslint-plugin": "^4.31.0",
     "@typescript-eslint/parser": "^4.31.0",
     "babel-jest": "^27.1.0",
-    "chokidar": "^3.5.2",
     "codemirror": "^5.62.3",
     "d3-dag": "^0.6.3",
     "d3-shape": "^3.0.1",
@@ -67,7 +66,6 @@
     "js-yaml": "^4.1.0",
     "livereload": "^0.9.3",
     "lodash": "^4.17.21",
-    "polka": "^1.0.0-next.20",
     "prettier": "^2.3.2",
     "prettier-plugin-svelte": "^2.4.0",
     "remark-parse": "10.0.0",
