+++ packages/viewer/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@irydium/viewer",
   "version": "0.0.14",
-  "description": "Live viewer for irydium files",
+  "description": "Live viewer for Irydium",
   "author": "William Lachance <wlach@protonmail.com>",
   "homepage": "https://github.com/irydium/irydium/tree/master/packages/viewer#readme",
   "license": "MIT",
@@ -30,7 +30,12 @@
   },
   "dependencies": {
     "@irydium/compiler": "workspace:^0.0.14",
+    "chokidar": "^3.5.2",
+    "commander": "^9.0.0",
     "livereload": "^0.9.1",
-    "polka": "next"
+    "mustache": "^4.0.1",
+    "polka": "next",
+    "sirv": "^1.0.0",
+    "ws": "^8.4.2"
   }
 }
